Rename checkpoints2 -> checkpoints
https://github.com/MystenLabs/sui/issues/5763
