+++ crates/sui-core/src/authority.rs
@@ -81,7 +81,7 @@ use sui_types::{
 
 use crate::authority::authority_notifier::TransactionNotifierTicket;
 use crate::authority::authority_notify_read::NotifyRead;
-use crate::checkpoints2::{CheckpointMetrics, CheckpointService, LogCheckpointOutput};
+use crate::checkpoints::{CheckpointMetrics, CheckpointService, LogCheckpointOutput};
 use crate::consensus_handler::{
     SequencedConsensusTransaction, VerifiedSequencedConsensusTransaction,
 };
