+++ crates/sui-core/src/lib.rs
@@ -23,7 +23,7 @@ pub mod validator_info;
 
 pub mod test_utils;
 
-pub mod checkpoints2;
+pub mod checkpoints;
 mod consensus_handler;
 mod histogram;
 mod node_sync;
